+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming 
 * (please add here)
 
+## 1.1.4 
+Added API call to hook to return connector schemas
+
 ## [1.1.3](https://github.com/fivetran/airflow-provider-fivetran/releases/tag/v1.1.3) - 2022-12-12
 Added ability to test a Fivetran Connection type